Provide PSGI wrapper for perlcritic.com
Map all the static files to the correct locations, and wrap the
perl/critic.pl CGI script with CGI::Emulate::PSGI. Also redirect the
base URL to index.html and find the /tmp location that corresponds to
what perl/critic.pl is using.

With this commit perlcritic.com can be tested locally with `plackup`.